fix tests, lint move percentage-bar to the right location
