+++ src/app/core-ui/core-ui.module.ts
@@ -1,4 +1,4 @@
-import { NgModule, ModuleWithProviders } from '@angular/core';
+import { NgModule, ModuleWithProviders, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 
@@ -30,7 +30,8 @@ import { PaginatorComponent } from './paginator/paginator.component';
     DirectivesModule,
     MaterialModule,
     PaginatorComponent
-  ]
+  ],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })
 export class CoreUiModule {
   static forRoot(): ModuleWithProviders {
