+++ src/app/installer/create-wallet/passphrase/passphrase.component.spec.ts
@@ -3,16 +3,17 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MatDialogRef, MatIconModule } from '@angular/material';
 
-import { ModalsModule, PassphraseService } from '../../modals.module';
+import { PassphraseService } from './passphrase.service';
 
 /* modules (own) */
 import { SharedModule } from '../../../wallet/shared/shared.module';
 import { CoreModule } from '../../../core/core.module';
 import { CoreUiModule } from '../../../core-ui/core-ui.module';
-
+import { InstallerModule } from 'app/installer/installer.module';
 /* components & directives (own) */
-import { FocusDirective } from '../../modals.directives';
+
 import { PassphraseComponent } from './passphrase.component';
+
 
 /* delete any unused imports! */
 
@@ -27,16 +28,15 @@ describe('PassphraseComponent', () => {
         SharedModule,
         CoreModule.forRoot(),
         CoreUiModule.forRoot(),
-        ModalsModule.forRoot(),
+        InstallerModule,
         /* deps */
         BrowserAnimationsModule,
-        MatIconModule,
+        // MatIconModule,
        ],
       providers: [
         /* own */
         PassphraseService,
         /* deps */
-        { provide: MatDialogRef},
       ]
     })
     .compileComponents();
