+++ src/app/installer/create-wallet/passphrase/passphrase.service.spec.ts
@@ -2,15 +2,14 @@ import { TestBed, inject } from '@angular/core/testing';
 
 import { SharedModule } from '../../../wallet/shared/shared.module';
 import { CoreModule} from '../../../core/core.module';
-import { ModalsModule, PassphraseService } from '../../modals.module';
+import { PassphraseService } from './passphrase.service';
 
 describe('PassphraseService', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [
         SharedModule,
-        CoreModule.forRoot(),
-        ModalsModule.forRoot()
+        CoreModule.forRoot()
       ],
       providers: [PassphraseService]
     });
