+++ src/app/installer/encrypt-wallet/encrypt-wallet.component.spec.ts
@@ -1,6 +1,13 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { RouterTestingModule } from '@angular/router/testing';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+
+import { CoreModule } from 'app/core/core.module';
+import { CoreUiModule } from 'app/core-ui/core-ui.module';
+import { InstallerModule } from 'app/installer/installer.module';
 
 import { EncryptWalletComponent } from './encrypt-wallet.component';
+
 
 describe('EncryptWalletComponent', () => {
   let component: EncryptWalletComponent;
@@ -8,7 +15,14 @@ describe('EncryptWalletComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ EncryptWalletComponent ]
+      // declarations: [ EncryptWalletComponent ]
+      imports: [
+        CoreModule.forRoot(),
+        CoreUiModule.forRoot(),
+        InstallerModule,
+        RouterTestingModule,
+        BrowserAnimationsModule
+      ]
     })
     .compileComponents();
   }));
