+++ src/app/installer/encrypt-wallet/encrypt-wallet.service.spec.ts
@@ -1,11 +1,19 @@
 import { TestBed, inject } from '@angular/core/testing';
 
+import { CoreModule } from 'app/core/core.module';
+import { RouterTestingModule } from '@angular/router/testing';
+
 import { EncryptWalletService } from './encrypt-wallet.service';
+
 
 describe('EncryptWalletService', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
-      providers: [EncryptWalletService]
+      providers: [EncryptWalletService],
+      imports: [
+        CoreModule.forRoot(),
+        RouterTestingModule,
+      ]
     });
   });
