+++ src/app/installer/installer.component.spec.ts
@@ -1,4 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { RouterTestingModule } from '@angular/router/testing';
+
+import { InstallerModule } from 'app/installer/installer.module';
 
 import { InstallerComponent } from './installer.component';
 
@@ -8,7 +11,10 @@ describe('InstallerComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ InstallerComponent ]
+      imports: [
+        InstallerModule,
+        RouterTestingModule,
+      ]
     })
     .compileComponents();
   }));
