+++ src/app/loading/connection-checker.service.spec.ts
@@ -1,11 +1,16 @@
 import { TestBed, inject } from '@angular/core/testing';
 
+import { CoreModule } from 'app/core/core.module';
+
 import { ConnectionCheckerService } from './connection-checker.service';
 
 describe('ConnectionCheckerService', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
-      providers: [ConnectionCheckerService]
+      providers: [ConnectionCheckerService],
+      imports: [
+        CoreModule.forRoot()
+      ]
     });
   });
