+++ src/app/loading/loading.component.spec.ts
@@ -1,4 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { RouterTestingModule } from '@angular/router/testing';
+
+import { CoreModule } from 'app/core/core.module';
 
 import { LoadingComponent } from './loading.component';
 
@@ -8,7 +11,11 @@ describe('LoadingComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ LoadingComponent ]
+      declarations: [ LoadingComponent ],
+      imports: [
+        CoreModule.forRoot(),
+        RouterTestingModule
+      ]
     })
     .compileComponents();
   }));
