+++ src/app/wallet/main/percentage-bar/percentage-bar.component.ts
